moved and removed files
