+++ public/js/components/specific/card/card-element.js
@@ -1,5 +1,5 @@
 import { LitElement, css, html } from 'lit-element';
-import { Card } from '../../pages/cards/card.js';
+import { Card } from '../../../pages/cards/card.js';
 export class CardElement extends LitElement {
 
     static get properties() {
