+++ public/js/pages/cards/index.js
@@ -1,5 +1,5 @@
-import '../../components/card/cardSwitcher.js';
-import '../../components/homeButton.js';
+import '../../components/global/homeButton.js';
+import '../../components/specific/card/cardSwitcher.js';
 
 import { getTheme } from '../../utils/operations.js';
 getTheme()
