+++ public/js/pages/home/index.js
@@ -1,5 +1,5 @@
-import '../../components/profile/profilesHolder.js';
+import '../../components/specific/profile/profilesHolder.js';
 
-import '../../components/themeSwitcher.js';
+import '../../components/global/themeSwitcher.js';
 
 document.querySelector('theme-switcher').setOptions(['default', 'night-fade', 'deep-blue', 'chrome', 'dense-waters', 'moon', 'sun', 'fresh'])
