+++ public/js/pages/practice/index.js
@@ -1,5 +1,5 @@
-import '../../components/homeButton.js';
-import '../../components/practice/cardViewer.js';
+import '../../components/global/homeButton.js';
+import '../../components/specific/practice/cardViewer.js';
 
 import { getTheme } from '../../utils/operations.js';
 getTheme()
