+++ public/views/index.html
@@ -7,8 +7,6 @@
     <link rel="stylesheet" href="/styles/theme.css">
     <script type="module" src="../js/pages/home/index.js"></script>
     <link rel="stylesheet" href="/styles/home/components/profiles.css">
-    <link rel="stylesheet" href="/styles/home/style.css">
-    <link rel="stylesheet" href="/styles/home/utils.css">
     <link rel="stylesheet" href="/styles/utils.css">
     <script type="text/javascript" src="https://livejs.com/live.js"></script> <!-- live server -->
     <title>Main page</title>
